- Upgrade to Java 17 - Run unit tests with Mongo 6.x and Mongo 7.x - Move common test configuration to a test module. Move logs to be under target and log4j.xml file to test module - Move MongoId annotation to core.  Remove unnecessary duplicate classes - Since QueryProvider is the same for Reactive and Sync operations, move it to cor eand use the same provider for both modules - Removed unnecessary nesting of parent projects since the same Spring libraries provide support for Reactive and Sync execution.  The only difference is in the Mongo drivers and the project structure now correctly reflects that difference - Successful migration to Spring 6.0 and Mongo 7.0 - Fixed NonReactiveQueryProvider

+++ mongodb-aggregate-query-support-core/src/main/java/com/github/krr/mongodb/aggregate/support/pageable/PageableFacet.java
@@ -34,16 +34,15 @@ public class PageableFacet implements Facet {
 
   private final int order;
 
-  private final int offset;
+  private final long offset;
 
-  private final int pageSize;
+  private final long pageSize;
 
-  public PageableFacet(int order, int offset, int pageSize) {
+  public PageableFacet(int order, long offset, long pageSize) {
     this.order = order;
     this.offset = offset;
     this.pageSize = pageSize;
   }
-
 
   @Override
   public Class<? extends Annotation> annotationType() {
